Allocate examples tasks based on an average execution time
Currently tasks are structured according to how they are created, which can result in uneven execution times, or in the case of tasks with a small number of examples, only a single task being created, missing the opportunity for parallel execution.

Add GtRrAverageTimeExampleAllocator which spreads example execution across multiple tasks, even for a small number of examples, and balances the number of examples executed in each task.

Fixes: feenkcom/gtoolkit#2591