PlpoGtRemoteRunner>>stop just stop the PharoLinkApplication
We should be telling the remote to shut down, however in some circumstances it is terminating the listen process and shutting down existing connections, but not getting to the point of shutting down the server socket.
Also, we should stop the PharoLinkApplication from listening for incoming messages since the connection is about to be closed.